Apply IntelliJ inspection cleanups; finish TimeoutOverride try-with-resources
Inspection pass over main, test fixtures, and both test suites:
FunctionRepeater.runUntilNotNull and FunctionRunner.run no longer
declare InterruptedException they never threw, removing ~25 dead
catch blocks across operators and drivers; redundant public/static/
final modifiers dropped from interfaces; inner classes that never
touched their outer instance made static; fields made final where
possible; slf4j calls parameterized; needless toString()/String
concatenation cleaned up; StringComparators strict/case-insensitive
drop a null check that String.equals/equalsIgnoreCase already provide;
ClassReference(Class) no longer declares ClassNotFoundException, and
EventDispatcher's robot creation catch list shrinks to match;
driver helper methods in final classes tightened to private.

Also completes the TimeoutOverride try-with-resources conversion:
EventToolTest, JDialogOperatorTest, DeepMenuPushTest,
EditorScrollingInTabsTest (inspection pass) plus WindowManagerJobsTest,
TabbedPanePageSwitchTest, MenuNavigationTest (outer override and the
three-way override as one multi-resource try), and the three @Disabled
JFileChooserOperatorTest dialog tests. The @BeforeEach/@AfterEach
field-held overrides stay as explicit cancel() - they must span the
whole test method - and TimeoutsTest keeps raw calls because it tests
the override mechanism itself.

Verified: NullAway still gates compileJava (canary check), unit suite
green, all UI test classes touched by this change green individually.

+++ src/userInterfaceTest/java/org/netbeans/jemmy/testing/JSliderScrollModelsTest.java
@@ -58,7 +58,7 @@ class JSliderScrollModelsTest {
             op.setScrollModel(JSliderOperator.CLICK_SCROLL_MODEL);
             op.scrollToValue(value);
             assertThat(Integer.parseInt(label.getText())).isEqualTo(value);
-            value = div10(range / 3);
+            value = div10((double) range / 3);
             op.setScrollModel(JSliderOperator.PUSH_AND_WAIT_SCROLL_MODEL);
             op.scrollToValue(value);
             assertThat(Integer.parseInt(label.getText())).isEqualTo(value);
